+++ sso/src/routes/usersRoutes.ts
@@ -4,7 +4,7 @@ import { parse, validate } from '../middleware/validate'
 import { z } from '../openapi/zod'
 import { listUsers } from '../controllers/users/list'
 import { usersListSchema } from '../schemas/users/usersListSchema'
-import { UserRole, userUpdateSchema, validateSchema } from '../schemas'
+import { userUpdateSchema, validateSchema } from '../schemas'
 import { getMe } from '../controllers/users/getMe'
 import { authenticate } from '../middleware/authenticate'
 import { updateUser } from '../controllers/users/update'
@@ -35,6 +35,6 @@ usersRoutes.patch(
     z.object({ params: z.object({ id: userIdSchema }), body: userUpdateSchema })
   ),
   authenticate,
-  authorize(UserRole.ADMIN),
+  authorize('ADMIN'),
   updateUser
 )